requirements.txt
pygame==2.5.2
imageio==2.35.1
discord.py==2.4.0